refactor: modularize architecture (services layer, schema split)
