+++ src/app/api/v1/candidates/[id]/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { db, candidates, NewCandidate } from "@/db";
 import { eq } from "drizzle-orm";
-import { requireAuth } from "@/lib/api-auth";
+import { requireAuth } from "@/services/auth";
 
 type Params = { params: Promise<{ id: string }> };
