+++ src/app/api/v1/news/announcements/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { db, announcements, NewAnnouncement } from "@/db";
 import { eq, desc, and, SQL } from "drizzle-orm";
-import { requireAuth, parsePaginationParams } from "@/lib/api-auth";
+import { requireAuth, parsePaginationParams } from "@/services/auth";
 
 // GET /api/v1/news/announcements - List announcements
 export async function GET(request: NextRequest) {
