+++ src/app/api/v1/news/curated/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { db, curatedLinks, NewCuratedLink } from "@/db";
 import { eq, desc, and, SQL } from "drizzle-orm";
-import { requireAuth, parsePaginationParams } from "@/lib/api-auth";
+import { requireAuth, parsePaginationParams } from "@/services/auth";
 
 // GET /api/v1/news/curated - List curated links
 export async function GET(request: NextRequest) {
