+++ src/db/schema/index.ts
@@ -1,0 +1,6 @@
+export * from "./jobs";
+export * from "./candidates";
+export * from "./news";
+export * from "./investments";
+export * from "./blog";
+export * from "./misc";
